Add __main__.py to make package executable.
